Routes and client and server components
