+++ firstnextproject/app/mainpage/notification/page.tsx
@@ -1,73 +1,12 @@
-'use client';
+import NotificationComponent from "@/app/components/AcceptCoffeComponents/NotificationComponent";
 
-import { useState, useEffect } from 'react';
-import AcceptCoffeComponent from '@/app/components/AcceptCoffeComponents/AcceptCoffeComponent';
 
-export interface Event {
-  eventId: string;
-  firstName: string;
-  lastName: string;
-  status: string;
-}
 
-export default function Brew() {
-  const [currentEvents, setCurrentEvents] = useState<Event[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
 
-  useEffect(() => {
-    const fetchEvents = async () => {
-      try {
-        const user = JSON.parse(localStorage.getItem('user') || '{}');
-        const token = localStorage.getItem('token');
-
-        if (!user.userId || !token) {
-          throw new Error('User or token not found');
-        }
-
-        const url = `http://46.101.127.179:8080/api/events/pending/${user.userId}`;
-        const response = await fetch(url, {
-          headers: {
-            "Authorization": `Bearer ${token}`,
-            "Content-Type": "application/json"
-          }
-        });
-
-        if (!response.ok) {
-          throw new Error('Failed to fetch events');
-        }
-
-        const data: Event[] = await response.json();
-        setCurrentEvents(data);
-      } catch (err) {
-        setError(err instanceof Error ? err.message : 'An error occurred');
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    fetchEvents();
-  }, []);
-
-  if (isLoading) return <div>Loading...</div>;
-  if (error) return <div>Error: {error}</div>;
-  if(currentEvents.length === 0) {
-    return(
-      <div className='flex items-center justify-center h-screen'>
-        <p className='text-lg text-red-600 font-semibold bg-red-100 p-4 rounded-md border border-red-300 shadow-md'>
-          You dont have any notifications, try to refresh the page
-        </p>
-      </div>
-    );
-  }
+export default function Notification() {
+  
 
   return (
-    <div className="flex items-center justify-center h-screen">
-      {currentEvents.map(event => (
-        <div key={event.eventId}>
-          <AcceptCoffeComponent event={event} />
-        </div>
-      ))}
-    </div>
+    <NotificationComponent />
   );
 }
